Search results

+++ db/migrate/20140227221608_create_spaces.rb
@@ -6,7 +6,7 @@ class CreateSpaces < ActiveRecord::Migration
       t.text :description
       t.string :dimensions
       t.text :ammenities
-      t.float :rate
+      t.integer :rate
       t.string :address
       t.string :city
       t.string :state
